initial commit of slack-dapp prototype
Disclaimer:
This is a prototype and not a finished product! Parts of the code were written very fast, just to test how smart contracts work. There is still a lot of room for improvement.
This prototype was developed for demonstration purposes. It was developed with the purpose to show people how smart contracts work. This code is not production ready. This especially accounts for the lottery contract. This contract was not tested enough to be used as real smart contract. In the current version it is to be used for demonstration purposes only!

Included in this commit:
* NodeJs server application that includes the slack-bot listener and the web3 API to communicate with a running Ethereum client (tested with geth)
* lottery contract written in Solidity
* some simple tests for the contract (most of them written with Dapple, others to be added)

Note that this prototype is not base on an Ethereum development framework (such as Truffle or Embark). Framework integration might be added in the future.